Remove 2 GB file limit.
Case 71041: Archive::Tar::Builder was unable to handle files larger
than 2 GB due to certain signed 32-bit values rolling over and
becoming negative. This commit switches the relevant counters to
off_t, which is 64-bit regardless of architecture because we build
with -D_FILE_OFFSET_BITS=64 set.

A test script has been added which verifies the fix.

+++ src/b_buffer.c
@@ -72,9 +72,9 @@ static inline size_t padded_size(size_t size) {
     return size + (B_BUFFER_BLOCK_SIZE - (size % B_BUFFER_BLOCK_SIZE));
 }
 
-ssize_t b_buffer_reclaim(b_buffer *buf, size_t used, size_t given) {
+off_t b_buffer_reclaim(b_buffer *buf, size_t used, size_t given) {
     size_t padded_len = padded_size(used);
-    ssize_t amount;
+    off_t amount;
 
     if (buf == NULL || given == 0 || given % B_BUFFER_BLOCK_SIZE) {
         errno = EINVAL;
