+++ src/b_file.h
@@ -7,7 +7,7 @@
 #include "b_string.h"
 #include "b_buffer.h"
 
-ssize_t b_file_write_contents(b_buffer *buf, int file_fd);
-ssize_t b_file_write_path_blocks(b_buffer *buf, b_string *path);
+off_t b_file_write_contents(b_buffer *buf, int file_fd);
+off_t b_file_write_path_blocks(b_buffer *buf, b_string *path);
 
 #endif /* _B_FILE_H */
